stabilize ephemeral-raid restart/start/stop methods
NOTE: testing only on CentOS so far

After finding that a start/start routine basically makes a mess, I normalized the behavior to swallow errors where acceptable, and to also avoid doing what's not necessary.  The start/stop procedures will autodetect if a raid array exists with the same name and if it does, it'll softly rebuild it - meaning it'll basically say "reassemble with MY rules" instead of letting mdadm's abilities to munge up devices break all the health checks.

the force-reload method is very destructive and works as advertized; it will destroy what is there, and reprovision.

Also, included some notes in the defaults file; this explains that providing a /dev/md/ path for your device is much more consitent.